+++ packages/terra-date-time-picker/src/terra-dev-site/test/date-time-picker/DateTimePickerWithTimezoneDst.test.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import DateTimePickerTimezoneExampleTemplate from '../../common/DateTimePickerTimezoneExampleTemplate';
+
+const DateTimePickerExample = () => (
+  <DateTimePickerTimezoneExampleTemplate
+    value="2017-11-05T01:30:00"
+    initialTimeZone="America/Chicago"
+  />
+);
+
+export default DateTimePickerExample;