+++ packages/terra-date-picker/src/terra-dev-site/test/date-picker/calendar-filter/CalendarFilterFilterDates.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import moment from 'moment';
+import moment from 'moment-timezone';
 import CalendarFilter from '../../../../CalendarFilter';
 
 const isWeekdayOrToday = (date) => {