+++ packages/terra-date-picker/CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## Unreleased
 
+* Added
+  * Added initialTimeZone a private prop that gets passed down from terra-date-time-picker.
+
 * Changed
+  * Changed 'moment' dependency to 'moment-timezone'
   * Updated to support a peer dependency of react-intl v2-v5
 
 ## 4.53.0 - (November 24, 2020)